Refactor CMakeLists.txt for DrHook extensions
This commit standardises the approach to adding third party extensions (such as NVTX) to DrHook's CMakeLists.txt.

The new approach is to exclude the extensions folder from the initial fiat target creation, and then individually add the extension's sources, include directories, and external libraries. This allows for an easy pattern to follow for future extensions, which is entirely contained within one `if (HAVE_DR_HOOK_{feature})` statement.