Add ability for users to specify PAPI events in Dr Hook (#27)
* Add DR_HOOK support for HW counters with PAPI

* Make fiat-drhook-sanity a unit-test, refactor CMake and other cosmetic
cleanup

* Fix invalid arguments error to PAPI_list_events

* Fix DR_HOOK_SILENT option

* Change error messages to be more explicit

Fix typo

* Fix missing intent

* Add new debug print

This new print clarifies what the next set of prints are referring to.

It also slightly optimises the subsequent control flow.

* Add functionality for user specified PAPI events

This allows user to specify their own PAPI events (up to MAXNPAPICNTRS, defaulting to 4).
This is done with the DR_HOOK_PAPI_COUNTERS flag. If the flag is not specified, and PAPI
is enabled, then the following defaults will be chosen:
PAPI_TOT_CYC
PAPI_FP_OPS
PAPI_L1_DCA
PAPI_L2_DCM

If an invalid event is chosen, then DrHook will simply crash will an appropriate error message.

* Fix whitespacing

This commit removes all tabs, in favour of two spaces, in files touched by #26 and #27.

* Fix malloc -> malloc_drhook

* Fix safe_thread_num -> papi_safe_thread_num

* Add if (opt_papi) guards

While PAPI code was behind #ifdefs, opt_papi was only used for toggling the printing of counters. PAPI would still be running as part of DrHook. To be inline with other options, PAPI is now behind both #ifdefs statically and if (opt_papi) dynamically.

* Refactor PAPI extension

* Fix NPAPICNTRS -> MAXNPAPICNTRS

* Remove static from malloc_drhook()

The qualifier needed to be removed so that malloc_drhook() could be used from within drhook_papi.c. To try and retain some restricted scoping (albeit no longer enforced by the compiler), malloc_drhook() hasn't been added to a header and extern is needed to use it.

* Move OML_MY_THREAD() into parallel region

Also fixed some related white spacing

* Add PAPI tests

* Fix nproc being initialised to 1

Previously nproc was initialised to 0. This meant that get_mon_out() wouldn't work as expected when not using MPI due to a bad assumption in an if statement.

* Rename dr_hook -> drhook

* Add copyright headers

* Revert "Fix nproc being initialised to 1"

This reverts commit 8a558aa205ee39cba6e72037eb799831eadf4dba.

* Fix output formatting

When outputting the value of DR_HOOK_PAPI_COUNTERS, the info prefix was missing.

* Change initialising process info

dr_hook_procinfo() now separates the setting of myproc & nproc from the reporting of initialising MPI. It will now initialise both myproc and nproc to 1 if DrHook is not running with MPI. To report the state of MPI, the new parameter mpi_init will be used.

* Fix drhook_papi_max_name_len() missing from header

* Add drhook_papi_user_counters_more_than_max test

Tests that any counters over MAXNPAPICNTRS are silently dropped.

* Fix set_tests_properties() not finding MPI tests

Previously set_tests_properties() would fail because it couldn't find  fiat_test_drhook_papi_mpi_valid_csv to add properties to. This was because fiat_test_drhook_papi_mpi_valid_csv was behind a condition that required MPI, while set_tests_properties() was not.

* Use tid argument vs drhook_oml_get_thread_num()

* Add guards to set_tests_properties() in tests

Prevents errors of not finding a test when trying to set properties.

* Add C_DR_HOOK_PROCINFO()

Used to ensure type safety between C and Fortran for DR_HOOK_PROCINFO()

* Fix compiler warnings

* Fix valid_csv tests

Tests were accidentally searching for the wrong string in the regex check

* Remove reliance on papi.h when building without PAPI

Previously papi.h was needed by drhook_papi.h, even when building without PAPI. Now it is only needed when PAPI is requested as the type is defined locally.

* Merge develop

---------

Co-authored-by: Martyn Foster <martyn.foster@gmail.com>
Co-authored-by: Willem Deconinck <willem.deconinck@ecmwf.int>